Add `TxBuilder::get_available_balances`
Besides the changes in the `TxBuilder` API, this is a code move.
